querygrid: benchmark tests module for querygrid
Summary:
Affter we commit our new changes to the test framework probably this module will not be actual.
But anyway, this pach contains several fixes, and the TPCH queryries for querygrid.
Just to don't lose it i would store it in the repository.

Task: SWARM-765
Review Url: @@review_url@@

Test Plan: manual on the local/perf clusters

Reviewers: sogorkis, gkokosinski

Reviewed By: sogorkis, gkokosinski

Differential Revision: https://phabricator.td.teradata.com/D834