Repeat query combinations instead of every query
This changes the order of execution of benchmark queries, where instead
of immediately repeating every query, all queries for a combination of
variables of a benchmark are repeated in a sequence. This includes
warmup queries.

Throughput tests execution order is not affected.

If there's an unexpected exception, results for all combinations of a
benchmark are annotated with it.

The main idea is to group benchmark combinations and pass them down so
they're executed in the same executor. Lots of methods now accept a
List<Benchmark>, where it can be a single item or a group. There are a
few new assertions to make sure that parameters like runs and
concurrency have the same value for all benchmarks in a group.

When executions are returned, they're grouped by benchmark again to
create results.